Corrected codescan pattern array size calculation.
I was basing it on having three pieces of information per OpCode, but it is actually four.  I think it was originally three, and that is what I based the array size on, but then I changed the passed information and never updated the other code since the only time that was an issue was in very long patterns.