Totally removed jump targets.
They made the stack levels FRACTIONALLY more accurate - if you have a match right at a `goto` target, before the stack has been fixed, then this new method will not reflect the stack size before the jump.  HOWEVER, this is not a problem:  previously this case would favour the stack size at the jump location, so if the stack was different between the source of the jump and the code immediately before the target, the former would be used.  There is no reason to choose one over the other - either way, it is indeterminate, so why not use the value that is VASTLY easier to get?